LayoutTests/fast/js/toString-try-else.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<link rel="stylesheet" href="resources/js-test-style.css">
<script src="resources/js-test-pre.js"></script>
</head>
<body>
<script>

    description(
    "This test checks that toString() round-trip on a function that has a else clause directly \
    followed by an else clause puts whitespace between the two on decompilation."
    );


    function testTryElse() { if (x) {} else try {} finally {} }

    unevalf = function(x) { return '(' + x.toString() + ')'; };
    
    // Test round trip.
    shouldBe("unevalf(eval(unevalf(testTryElse)))", "unevalf(testTryElse)");

    var successfullyParsed = true;
</script>
<script src="resources/js-test-post.js"></script>
</body>
</html>
